requirements_freeze.txt
absl-py==2.3.1
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.4.0
beartype==0.22.5
beautifulsoup4==4.14.2
biopandas==0.5.1
biopython==1.86
bioservices==1.12.1
bitsandbytes==0.48.2
cattrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.2
cosine_annealing_warmup @ git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup@12d03c07553aedd3d9e9155e2b3e31ce8c64081a
cpdb-protein==0.2.0
cycler==0.12.1
Cython==3.2.0
deepdiff==8.6.1
easydev==0.13.3
einops==0.8.1
einx==0.3.0
exceptiongroup==1.3.0
filelock==3.19.1
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.9.0
gevent==25.9.1
graphein==1.7.7
greenlet==3.2.4
grequests==0.7.0
grpcio==1.76.0
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
jaxtyping==0.3.3
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lightning-utilities==0.15.2
line_profiler==4.2.0
loguru==0.7.3
looseversion==1.1.2
lxml==5.4.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
mmtf-python==1.1.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multipledispatch==1.0.0
narwhals==2.11.0
ndlinear==1.0.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
orderly-set==5.5.0
packaging==25.0
pandas==2.3.3
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
plotly==6.4.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-box==7.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-cache==1.2.1
rich==14.2.0
rich-click==1.9.4
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
soupsieve==2.8
suds-community==1.2.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==1.0.22
tmtools==0.3.0
tokenizers==0.22.1
torch==2.5.1+cu121
torch-geometric==2.7.0
torch-tb-profiler==0.4.3
torch_cluster==1.6.3+pt25cu121
torch_scatter==2.1.2+pt25cu121
torchaudio==2.5.1+cu121
torchmetrics==1.8.2
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.1
triton==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
vector-quantize-pytorch==1.25.1
wadler_lindig==0.1.7
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.3
x-transformers==2.11.23
xarray==2025.6.1
xmltodict==0.14.2
xxhash==3.6.0
yarl==1.22.0
zope.event==6.1
zope.interface==8.1
